# Reduce layout thrashing by optimizing dxSTable row creation