--- a/js/webui.js
+++ b/js/webui.js
@@ -1633,7 +1633,8 @@ var theWebUI =
    		var tul = 0;
 		var tdl = 0;
 		var tArray = [];
-		var tMaxRows = Math.ceil(Math.min(table.dBody.clientHeight,table.dCont.clientHeight) / table.trHeight);
+		var firstLoad = this.firstLoad;
+		
 		$.each(data.torrents,
 		/**
 		 * @param {string} hash - torrent hash
@@ -1649,7 +1650,7 @@ var theWebUI =
 			if(!$type(theWebUI.torrents[hash]))
 			{
 				theWebUI.labels[hash] = lbl;
-				table.addRowById(torrent, hash, sInfo[0], {label : lbl}, tMaxRows);
+				table.addRowById(torrent, hash, sInfo[0], {label : lbl}, firstLoad);
 				tArray.push(hash);
 				theWebUI.filterByLabel(hash);
 			}
@@ -1748,8 +1749,8 @@ var theWebUI =
 		else 
 		{
 			table.refreshRows();
+			table.Sort();
 		}
-		table.Sort();
 		this.setInterval();
 		this.updateDetails();
    	},