@Novik 2021-12-20
Thanks, but IMHO instead this patch we need to optimize function dxSTable.prototype.getMaxRows (for example, remove condition 'this.viewRows<this.maxViewRows').
Now we haven't ascient IE, Safari etc.
I will check it in later.