@stickz 2021-12-21
Do you think it's a good idea to rewrite this function though? There are other methods which call `dxSTable.prototype.getMaxRows`. This could have unintended side effects when calling them.

I investigated your suggestion further and found that `dxSTable.prototype.refreshRows` also calls this function afterwards, to correct the hack to the reflow issue I created. That is not going to work properly, without further complicating the code.