@stickz 2021-12-23
@Novik Could you review this pull request again? I optimized dxSTable row creation and removed an unnecessary table sort. The results are even more astonishing than before and the code quality is much higher.

**Benchmarks for theWebUI.addTorrents()**
Loading times with 500 torrents are reduced from 800ms to 175ms. Approximately 70ms is a performance penalty.
Loading times with 5000 torrents are reduced from 18s to 520ms. Approximately 70ms is a performance penalty.

The **performance penalty** is now reduced to approximately 70ms, regardless of the number of torrents.