@Novik 2021-12-23
> Loading times with 5000 torrents are reduced from 18s to 520ms. 

And this is not surprising, because really you simple remove DOM filling at the first call of function. I.e. if we will have only one call of addTorrents(), then user will see nothing.
I think in the such way you may reduce timings yet more - if you write something like if(this.firstLoad) { this.show(); return; } at the start of function...
Can you explain your changes, please?