@stickz 2021-12-23
> And this is not surprising, because really you simple remove DOM filling at the first call of function. I.e. if we will have only one call of addTorrents(), then user will see nothing. 

`loadTorrents()` is called inside the `addTorrents()` function. This goes through and fills the DOM after dxSTable rowIDs Array is populated. This improves the flow so the code doesn't thrash the layout. We still have to populate this array to fill the DOM.