@Novik 2021-12-23
> This goes through and fills the DOM after dxSTable rowIDs Array is populated.

Yes, you are right, through the sequence loadTorrents -> show -> resize -> resizeTop -> dxSTable.resize -> dxSTable.refreshRows.

But yet one moment. Pay attention to this fragment, please:

```
        if (!fast)
	{
             ....
	}
	else
	{
             ....
	}

	var self = this;
	if((this.sIndex !=- 1) && !this.noSort)
		this.sortTimeout = window.setTimeout(function() { self.Sort(); }, 200);
```

We may have here sIndex !=- 1 (if user press on the column header in the previous session, for example).
As result, we will have call of Sort. We really need this in the "fast" case?