@stickz 2021-12-23
> We really need this in the "fast" case?
Good catch. I just removed the table sort from the "fast" case. It's only taking about 300ms now to add the torrents, regardless of the number of torrents added to the web client. I think this PR is ready for merge now.